Preprocessing+splitting
I did some data cleaning by removing the NaN values where necessary.
Created an extra column that is purely string and will be used as a feature of the data to be predicted.
Did some conversions from series or dataframe columns to numpy arrays as the tensorflow models understands arrays. Did some label encoding as the tensorflow model can predict numeric values only. And did the splitting of the features and the value to be predicted.